Add http status dog and add http_cat to group http_status
